AMD64 - Sync AMD64 support from Jordan Gordeev's svn repository and Google SOC project.  This work is still continuing but represents substantial progress in the effort.
With this commit the world builds and installs, the loader is able to
boot the kernel, and the kernel is able to initialize, probe devices, and
exec the init program.  The init program is then able to run until it hits
its first fork().  For the purposes of the GSOC the project is being
considered a big success!

The code has been adapted from multiple sources, most notably Peter Wemm
and other peoples work from FreeBSD, with many modifications to make it
work with DragonFly.  Also thanks go to Simon Schubert for working on gdb
and compiler issues, and to Noah Yan for a good chunk of precursor work
in 2007.

While Jordan wishes to be modest on his contribution, frankly we would
not have been able to make this much progress without the large number
of man-hours Jordan has dedicated to his GSOC project painstakingly gluing
code together, tracking down issues, and progressing the boot sequence.

Submitted-by: Jordan Gordeev <jgordeev@dir.bg>

+++ sys/sys/tls.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2005 The DragonFly Project.  All rights reserved.
+ * Copyright (c) 2005,2008 The DragonFly Project.  All rights reserved.
  * 
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -28,7 +28,7 @@
  * OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- * $DragonFly: src/sys/sys/tls.h,v 1.9 2008/08/25 23:34:35 dillon Exp $
+ * $DragonFly: src/sys/sys/tls.h,v 1.10 2008/08/29 17:07:04 dillon Exp $
  */
 
 #ifndef _SYS_TLS_H_
@@ -38,7 +38,7 @@
 
 struct tls_info {
 	void *base;
-	long size;
+	size_t size;
 };
 
 int set_tls_area(int which, struct tls_info *info, size_t infosize);
